+++ otelconf/config_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/stretchr/testify/require"
 	lognoop "go.opentelemetry.io/otel/log/noop"
 	metricnoop "go.opentelemetry.io/otel/metric/noop"
+	"go.opentelemetry.io/otel/propagation"
 	sdklog "go.opentelemetry.io/otel/sdk/log"
 	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
@@ -356,6 +357,7 @@ func TestNewSDK(t *testing.T) {
 		wantTracerProvider any
 		wantMeterProvider  any
 		wantLoggerProvider any
+		wantPropagator     any
 		wantErr            error
 		wantShutdownErr    error
 	}{
@@ -364,6 +366,7 @@ func TestNewSDK(t *testing.T) {
 			wantTracerProvider: tracenoop.NewTracerProvider(),
 			wantMeterProvider:  metricnoop.NewMeterProvider(),
 			wantLoggerProvider: lognoop.NewLoggerProvider(),
+			wantPropagator:     propagation.NewCompositeTextMapPropagator(),
 		},
 		{
 			name: "with-configuration",
@@ -373,11 +376,13 @@ func TestNewSDK(t *testing.T) {
 					TracerProvider: &TracerProviderJson{},
 					MeterProvider:  &MeterProviderJson{},
 					LoggerProvider: &LoggerProviderJson{},
+					Propagator:     &PropagatorJson{},
 				}),
 			},
 			wantTracerProvider: &sdktrace.TracerProvider{},
 			wantMeterProvider:  &sdkmetric.MeterProvider{},
 			wantLoggerProvider: &sdklog.LoggerProvider{},
+			wantPropagator:     propagation.NewCompositeTextMapPropagator(),
 		},
 		{
 			name: "with-sdk-disabled",
@@ -393,6 +398,7 @@ func TestNewSDK(t *testing.T) {
 			wantTracerProvider: tracenoop.NewTracerProvider(),
 			wantMeterProvider:  metricnoop.NewMeterProvider(),
 			wantLoggerProvider: lognoop.NewLoggerProvider(),
+			wantPropagator:     propagation.NewCompositeTextMapPropagator(),
 		},
 		{
 			name: "invalid resource",
@@ -409,6 +415,24 @@ func TestNewSDK(t *testing.T) {
 			wantTracerProvider: tracenoop.NewTracerProvider(),
 			wantMeterProvider:  metricnoop.NewMeterProvider(),
 			wantLoggerProvider: lognoop.NewLoggerProvider(),
+			wantPropagator:     propagation.NewCompositeTextMapPropagator(),
+		},
+		{
+			name: "invalid propagator",
+			cfg: []ConfigurationOption{
+				WithContext(t.Context()),
+				WithOpenTelemetryConfiguration(OpenTelemetryConfiguration{
+					TracerProvider: &TracerProviderJson{},
+					MeterProvider:  &MeterProviderJson{},
+					LoggerProvider: &LoggerProviderJson{},
+					Propagator:     &LoggerProviderJson{},
+				}),
+			},
+			wantErr:            newErrInvalid("propagator"),
+			wantTracerProvider: tracenoop.NewTracerProvider(),
+			wantMeterProvider:  metricnoop.NewMeterProvider(),
+			wantLoggerProvider: lognoop.NewLoggerProvider(),
+			wantPropagator:     propagation.NewCompositeTextMapPropagator(),
 		},
 		{
 			name: "invalid logger provider",
@@ -425,6 +449,7 @@ func TestNewSDK(t *testing.T) {
 			wantTracerProvider: tracenoop.NewTracerProvider(),
 			wantMeterProvider:  metricnoop.NewMeterProvider(),
 			wantLoggerProvider: lognoop.NewLoggerProvider(),
+			wantPropagator:     propagation.NewCompositeTextMapPropagator(),
 		},
 		{
 			name: "invalid tracer provider",
@@ -438,6 +463,7 @@ func TestNewSDK(t *testing.T) {
 			wantTracerProvider: tracenoop.NewTracerProvider(),
 			wantMeterProvider:  metricnoop.NewMeterProvider(),
 			wantLoggerProvider: lognoop.NewLoggerProvider(),
+			wantPropagator:     propagation.NewCompositeTextMapPropagator(),
 		},
 	}
 	for _, tt := range tests {
@@ -446,6 +472,7 @@ func TestNewSDK(t *testing.T) {
 		assert.IsType(t, tt.wantTracerProvider, sdk.TracerProvider())
 		assert.IsType(t, tt.wantMeterProvider, sdk.MeterProvider())
 		assert.IsType(t, tt.wantLoggerProvider, sdk.LoggerProvider())
+		assert.IsType(t, tt.wantPropagator, sdk.Propagator())
 		require.Equal(t, tt.wantShutdownErr, sdk.Shutdown(t.Context()))
 	}
 }